+++ src/components/layouts/Header/Menu.tsx
@@ -1,0 +1,14 @@
+'use client';
+
+import MenuIcon from '@/assets/icons/common/menu.svg';
+import { useMenuContext } from '@/context/MenuContext';
+
+export default function Menu() {
+  const { setIsOpen } = useMenuContext();
+
+  return (
+    <button onClick={() => setIsOpen(true)} className='cursor-pointer'>
+      <MenuIcon />
+    </button>
+  );
+}
